feat: use io/fs for searching files on Go 1.16+
